backend/app/utils/ai_integration.py
from deepseek import load_model  # Replace with your actual model loading code

model = load_model("path/to/your/deepseek_model")

def analyze_video(video_path: str):
    # Process video and return predictions (e.g., facial/motor analysis)
    predictions = model.predict_video(video_path)
    return {"status": "success", "predictions": predictions}

def analyze_audio(audio_path: str):
    # Process audio and return speech analysis
    predictions = model.predict_audio(audio_path)
    return {"status": "success", "predictions": predictions}

def chatbot_response(query: str):
    # Use DeepSeek as a chatbot
    response = model.generate_response(query)
    return {"response": response}