backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
firebase-admin>=6.0.1
python-dotenv>=0.19.0
python-multipart>=0.0.5